Added Input path and fix a few things

+++ delete_pals_save.py
@@ -1,17 +1,22 @@
-import re, os , sys
+import re, os , sys, glob
+def check_folder(directory, partial_text):
+    for filename in glob.glob(f"{directory}/*"):
+        if partial_text in filename:
+            return True
+    return False
 def get_number_in_range(min_value, max_value):
-  while True:
-    try:
-      number = int(input(f"Enter a number between {min_value} and {max_value}: "))
-      if min_value <= number <= max_value:
-        return number
-      else:
-        print("Number is out of range. Try again.")
-    except ValueError:
-      print("Invalid input. Please enter an integer.")
+    while True:
+        try:
+            number = int(input(f"Enter a number between {min_value} and {max_value}: "))
+            if min_value <= number <= max_value:
+                return number
+            else:
+                print("Number is out of range. Try again.")
+        except ValueError:
+            print("Invalid input. Please enter an integer.")
 def get_directory_from_user():
     while True:
-        directory_path = input("Please enter a directory path: ")
+        directory_path = input("Please enter a directory path of your players folder: ")
         if os.path.isdir(directory_path):
             return directory_path
         else:
@@ -50,24 +55,28 @@ def delete_player_saves(player_info):
         players_folder = get_directory_from_user()
     if not os.path.exists(players_folder):
         print(f"Players folder '{players_folder}' not found.")
-        return
-    total_pals_to_delete = sum(pals_found for _, _, pals_found in player_info)
-    deleted_count = 0
-    for line, uid, _ in player_info:
-        uid_no_hyphens = uid.replace('-', '')
-        sav_filename = f"{uid_no_hyphens}.sav"
-        sav_file_paths = [os.path.join(players_folder, f) for f in os.listdir(players_folder) if f.lower() == sav_filename.lower()]
-        if sav_file_paths:
-            sav_file_path = sav_file_paths[0]
-            os.remove(sav_file_path)
-            deleted_count += 1
-            print(f"Deleted: {sav_file_path}")
-            print(f"Deleted Player Info: {line}")
-    if deleted_count == 0:
-        print(f"No PlayerUID.sav files found for deletion, skipping...")
+        exit(1)
+    if check_folder(players_folder, "00"):
+        total_pals_to_delete = sum(pals_found for _, _, pals_found in player_info)
+        deleted_count = 0
+        for line, uid, _ in player_info:
+            uid_no_hyphens = uid.replace('-', '')
+            sav_filename = f"{uid_no_hyphens}.sav"
+            sav_file_paths = [os.path.join(players_folder, f) for f in os.listdir(players_folder) if f.lower() == sav_filename.lower()]
+            if sav_file_paths:
+                sav_file_path = sav_file_paths[0]
+                os.remove(sav_file_path)
+                deleted_count += 1
+                print(f"Deleted: {sav_file_path}")
+                print(f"Deleted Player Info: {line}")
+        if deleted_count == 0:
+            print(f"No PlayerUID.sav files found for deletion, skipping...")
+        else:
+            print(f"Total number of pals deleted: {total_pals_to_delete}")
+            print(f"Deleted {deleted_count} PlayerUID.sav files.")
     else:
-        print(f"Total number of pals deleted: {total_pals_to_delete}")
-        print(f"Deleted {deleted_count} PlayerUID.sav files.")
+        print(f"Players folder is empty or does not contain player sav files.")
+        exit(1)
 if __name__ == "__main__":
     if len(sys.argv) < 1:
         print("Usage: python delete_pals_save.py <log_file>")
